Rename webhook case files to kebab case
