+++ server/server/controllers/webhook/cases/story-moved-from-visible-column.js
